+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note: Can be used using nvuillam/mega-linter@insiders in your mega-linter.yml fi
 
 - Add link to linters rules index in documentation
 - Remove ANSI color codes from log files
+- Add performances by linter in console log
 
 - NPM package **Mega-Linter runner**
   - runs Mega-Linter locally, using .mega-linter.yml configuration (requires docker installed on your computer)